requirements.txt
aiohttp==3.13.3
beautifulsoup4==4.14.3
lxml==6.0.2
markdownify==1.2.2
python-dotenv==1.2.1
git+https://github.com/MCausc78/stoat.py@master
